LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/read-text/load-text-plain-expected.txt


FAIL Checking document metadata for text file assert_equals: expected "CSS1Compat" but got "BackCompat"
PASS Checking DOM for text file
PASS Checking contents for text file
